# marketplace/services/product-service/openapi/marketplace-api.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  description: API маркетплейса с товарами, заказами и промокодами
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Локальный сервер разработки

tags:
  - name: Products
    description: Управление каталогом товаров
  - name: Orders
    description: Управление заказами
  - name: Promo Codes
    description: Управление промокодами

paths:
  /products:
    post:
      tags:
        - Products
      summary: Создать новый товар
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Товар успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Products
      summary: Получить список товаров с пагинацией и фильтрацией
      operationId: getProducts
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Фильтр по статусу товара
          required: false
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          description: Фильтр по категории (точное совпадение)
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/{id}:
    get:
      tags:
        - Products
      summary: Получить товар по ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          description: ID товара
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали товара
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Products
      summary: Обновить товар
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          description: ID товара
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Товар успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Products
      summary: Мягкое удаление товара (архивация)
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          description: ID товара
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Товар успешно архивирован
        '404':
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders:
    post:
      tags:
        - Orders
      summary: Создать новый заказ
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Конфликт бизнес-логики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка промокода
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Превышен лимит частоты запросов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Получить заказ по ID
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          description: ID заказа
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Orders
      summary: Обновить заказ
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          description: ID заказа
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Заказ успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Конфликт бизнес-логики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Превышен лимит частоты запросов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Отменить заказ
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          description: ID заказа
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Заказ успешно отменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недопустимый переход состояния
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /promo-codes:
    post:
      tags:
        - Promo Codes
      summary: Создать новый промокод
      operationId: createPromoCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreate'
      responses:
        '201':
          description: Промокод успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ProductStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - ARCHIVED
      description: Статус товара

    ProductCreate:
      type: object
      required:
        - name
        - price
        - stock
        - category
        - status
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Название товара
        description:
          type: string
          maxLength: 4000
          nullable: true
          description: Описание товара
        price:
          type: number
          format: decimal
          minimum: 0.01
          exclusiveMinimum: true
          description: Цена товара (должна быть больше 0)
        stock:
          type: integer
          minimum: 0
          description: Количество на складе
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Категория товара
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Название товара
        description:
          type: string
          maxLength: 4000
          nullable: true
          description: Описание товара
        price:
          type: number
          format: decimal
          minimum: 0.01
          exclusiveMinimum: true
          description: Цена товара
        stock:
          type: integer
          minimum: 0
          description: Количество на складе
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Категория товара
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductResponse:
      type: object
      required:
        - id
        - name
        - price
        - stock
        - category
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: ID товара
        name:
          type: string
          description: Название товара
        description:
          type: string
          nullable: true
          description: Описание товара
        price:
          type: number
          format: decimal
          description: Цена товара
        stock:
          type: integer
          description: Количество на складе
        category:
          type: string
          description: Категория товара
        status:
          $ref: '#/components/schemas/ProductStatus'
        created_at:
          type: string
          format: date-time
          description: Время создания
        updated_at:
          type: string
          format: date-time
          description: Время последнего обновления

    ProductListResponse:
      type: object
      required:
        - items
        - total_elements
        - page
        - size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        total_elements:
          type: integer
          description: Общее количество товаров
        page:
          type: integer
          description: Номер текущей страницы
        size:
          type: integer
          description: Размер страницы

    OrderStatus:
      type: string
      enum:
        - CREATED
        - PAYMENT_PENDING
        - PAID
        - SHIPPED
        - COMPLETED
        - CANCELED
      description: Статус заказа

    OrderItemCreate:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
          format: uuid
          description: ID товара
        quantity:
          type: integer
          minimum: 1
          maximum: 999
          description: Количество товара

    OrderCreate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemCreate'
        promo_code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          nullable: true
          description: Промокод для применения

    OrderUpdate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemCreate'

    OrderItemResponse:
      type: object
      required:
        - id
        - product_id
        - quantity
        - price_at_order
      properties:
        id:
          type: string
          format: uuid
          description: ID позиции заказа
        product_id:
          type: string
          format: uuid
          description: ID товара
        quantity:
          type: integer
          description: Количество
        price_at_order:
          type: number
          format: decimal
          description: Цена на момент заказа

    OrderResponse:
      type: object
      required:
        - id
        - user_id
        - status
        - items
        - total_amount
        - discount_amount
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: ID заказа
        user_id:
          type: string
          format: uuid
          description: ID пользователя
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        promo_code_id:
          type: string
          format: uuid
          nullable: true
          description: ID примененного промокода
        total_amount:
          type: number
          format: decimal
          description: Итоговая сумма заказа
        discount_amount:
          type: number
          format: decimal
          description: Сумма скидки
        created_at:
          type: string
          format: date-time
          description: Время создания
        updated_at:
          type: string
          format: date-time
          description: Время последнего обновления

    DiscountType:
      type: string
      enum:
        - PERCENTAGE
        - FIXED_AMOUNT
      description: Тип скидки

    PromoCodeCreate:
      type: object
      required:
        - code
        - discount_type
        - discount_value
        - min_order_amount
        - max_uses
        - valid_from
        - valid_until
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          description: Промокод
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
          minimum: 0.01
          description: Значение скидки (процент или фиксированная сумма)
        min_order_amount:
          type: number
          format: decimal
          minimum: 0
          description: Минимальная сумма заказа для применения промокода
        max_uses:
          type: integer
          minimum: 1
          description: Максимальное количество использований
        valid_from:
          type: string
          format: date-time
          description: Начало срока действия
        valid_until:
          type: string
          format: date-time
          description: Окончание срока действия
        active:
          type: boolean
          default: true
          description: Активен ли промокод

    PromoCodeResponse:
      type: object
      required:
        - id
        - code
        - discount_type
        - discount_value
        - min_order_amount
        - max_uses
        - current_uses
        - valid_from
        - valid_until
        - active
      properties:
        id:
          type: string
          format: uuid
          description: ID промокода
        code:
          type: string
          description: Промокод
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
          description: Значение скидки
        min_order_amount:
          type: number
          format: decimal
          description: Минимальная сумма заказа
        max_uses:
          type: integer
          description: Максимальное количество использований
        current_uses:
          type: integer
          description: Текущее количество использований
        valid_from:
          type: string
          format: date-time
          description: Начало действия
        valid_until:
          type: string
          format: date-time
          description: Окончание действия
        active:
          type: boolean
          description: Активен ли промокод

    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          description: Машиночитаемый код ошибки
          enum:
            - PRODUCT_NOT_FOUND
            - PRODUCT_INACTIVE
            - ORDER_NOT_FOUND
            - ORDER_LIMIT_EXCEEDED
            - ORDER_HAS_ACTIVE
            - INVALID_STATE_TRANSITION
            - INSUFFICIENT_STOCK
            - PROMO_CODE_INVALID
            - PROMO_CODE_MIN_AMOUNT
            - ORDER_OWNERSHIP_VIOLATION
            - VALIDATION_ERROR
        message:
          type: string
          description: Человекочитаемое описание ошибки
        details:
          type: object
          nullable: true
          description: Дополнительные данные об ошибке
          additionalProperties: true